Change format for storing etcd machine address
Once the first etcd member is initialized, the machine controller has
to update the secret with the address of the machine, so it can be used
by other members to join during cluster creation.
The etcdadm-bootstrap-provider changes this address into an etcd client URL
before passing it in to the join command.
Recent changes in etcdadm, and etcdadm-controller will allow passing in
client URLs of all etcd members. So this commit changes the format of stored
address for the first machine from an IP address to the etcd client URL.
NOTE: This only happens once initially during cluster creation. We need to
keep this Secret because after clusterctl move, the etcdCluster's Initialized
condition needs to be set based on the existence of this Secret.